requirements.txt
pathlib
twitter
matplotlib
pytest
pytest-cov
config
tweepy
yweather
woeid
flask
Werkzeug
google
google-cloud-storage
gunicorn
flask-table
cachetools
tinydb
Flask-JSGlue
boto3
emoji-country-flag
mysql-connector-python
setuptools
Flask-Testing
flask-mysql